# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Progress API",
    "description": "GitHub multi-repo project tracking tool.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "tags": ["config"],
        "summary": "Get All Sections",
        "description": "Return ``{core, plugins}`` with secrets masked (spec 02).",
        "operationId": "get_all_sections_api_v1_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/config/language": {
      "get": {
        "tags": ["config"],
        "summary": "Get Language",
        "description": "Return the configured UI/notifications language (``core.language``).\n\nThe SPA fetches this on boot so its initial locale matches the server\nconfig, closing the gap where the web UI stayed English even though\n``core.language = zh-Hans`` was set (the SPA previously only honoured\nlocalStorage / the browser's Accept-Language).",
        "operationId": "get_language_api_v1_config_language_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": ["config"],
        "summary": "Set Language",
        "description": "Update ``core.language`` atomically (used by the language switcher).\n\nMerges the new value into the existing ``[core]`` section so the rest of\nthe core config (tokens, markpost url, …) is preserved, then refreshes\n``app.state.cfg`` so the LocaleMiddleware picks up the new language on the\nvery next request — no restart needed.",
        "operationId": "set_language_api_v1_config_language_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/config/schema": {
      "get": {
        "tags": ["config"],
        "summary": "Get Schema",
        "description": "Return per-section JSON Schemas for the frontend config editor.",
        "operationId": "get_schema_api_v1_config_schema_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSchemaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/config/{section}": {
      "put": {
        "tags": ["config"],
        "summary": "Put Section",
        "description": "Validate ``data`` against the section's schema and upsert it.",
        "operationId": "put_section_api_v1_config__section__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "section",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/config/reload": {
      "post": {
        "tags": ["config"],
        "summary": "Reload Config",
        "description": "Re-read DB-stored core config and refresh ``app.state.cfg``.",
        "operationId": "reload_config_api_v1_config_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReloadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/config/notifications/test": {
      "post": {
        "tags": ["config"],
        "summary": "Test Notifications",
        "description": "Send a test notification to all enabled channels; report per-channel results.\n\nUsed by the config console to verify channel reachability. Renders a fixed\ni18n test message (localized per the request's Accept-Language via\n``LocaleMiddleware``) and dispatches via the same ``Dispatcher`` the report\npipeline uses.",
        "operationId": "test_notifications_api_v1_config_notifications_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestNotificationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/integrations": {
      "get": {
        "tags": ["integrations"],
        "summary": "List Integrations",
        "description": "Registered integration list (spec 12).",
        "operationId": "list_integrations_api_v1_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationSummary"
                  },
                  "type": "array",
                  "title": "Response List Integrations Api V1 Integrations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/integrations/{name}/status": {
      "get": {
        "tags": ["integrations"],
        "summary": "Integration Status",
        "description": "Return one integration's status summary (spec 12).",
        "operationId": "integration_status_api_v1_integrations__name__status_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllConfigResponse": {
        "properties": {
          "core": {
            "additionalProperties": true,
            "type": "object",
            "title": "Core"
          },
          "plugins": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Plugins"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AllConfigResponse",
        "description": "``{section: data}`` map for ``GET /api/v1/config``."
      },
...
      "ConfigReloadResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ok"
          },
          "section": {
            "type": "string",
            "title": "Section",
            "default": "core"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConfigReloadResponse",
        "description": "Acknowledgement for ``POST /api/v1/config/reload``."
      },
      "ConfigSchemaResponse": {
        "properties": {
          "schemas": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Schemas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["schemas"],
        "title": "ConfigSchemaResponse",
        "description": "``{section: JSON Schema}`` map for ``GET /api/v1/config/schema``."
      },
      "ConfigSectionResponse": {
        "properties": {
          "section": {
            "type": "string",
            "title": "Section"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["section", "data"],
        "title": "ConfigSectionResponse",
        "description": "Per-section config dump with secrets masked (SecretStr → ``**********``)."
      },
      "ConfigUpdateRequest": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["data"],
        "title": "ConfigUpdateRequest",
        "description": "Body for ``PUT /api/v1/config/{section}``."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegrationSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "config_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Schema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["name"],
        "title": "IntegrationSummary",
        "description": "One registered integration in ``GET /api/v1/integrations``."
      },
      "LanguageResponse": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["language"],
        "title": "LanguageResponse",
        "description": "Configured UI/notifications language for ``GET /api/v1/config/language``.\n\nThe SPA reads this on boot so its initial locale matches the server's\n``core.language`` (instead of only honouring localStorage / the browser's\nAccept-Language, which left the UI in English even when zh-Hans was\nconfigured). Returned unmasked — language is not a secret."
      },
      "LanguageUpdateRequest": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["language"],
        "title": "LanguageUpdateRequest",
        "description": "Body for ``PUT /api/v1/config/language``.\n\nA dedicated single-field endpoint (rather than ``PUT /api/v1/config/core``)\nso the language switcher can change language atomically without having to\nround-trip the whole (masked) core section."
      },
...
      "TestChannelResult": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["channel", "ok"],
        "title": "TestChannelResult",
        "description": "Per-channel outcome of a test notification dispatch."
      },
      "TestNotificationResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/TestChannelResult"
            },
            "type": "array",
            "title": "Results"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["results", "summary"],
        "title": "TestNotificationResponse",
        "description": "Aggregated outcome of ``POST /api/v1/config/notifications/test``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}